+++ NEWS.md
@@ -1,3 +1,7 @@
 
-## OdsReaders.jl Trial 1 Release Note 2018-09-28
+## OdsReaders.jl v0.2.0 Release Note 2018-10-07
+* Add to searchods function
+* Add to searchsheet function
+
+## OdsReaders.jl v0.1.0 Release Note 2018-09-28
 * New Creation